requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
datasets>=2.20.0
openrouter>=0.7.11